# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Unauthorized" }
        },
        "summary": "Login with email and password",
        "tags": ["auth"]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [
          {
            "name": "invitationToken",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered successfully" },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized - Invalid invitation" }
        },
        "summary": "Register a new user account",
        "tags": ["auth"]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": { "description": "Returns the user profile" },
          "401": { "description": "Unauthorized" }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get the current user profile",
        "tags": ["auth"]
      }
    },
    "/auth/create-super-admin": {
      "post": {
        "operationId": "AuthController_createSuperAdmin",
        "parameters": [
          {
            "name": "secretKey",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Super admin created successfully" },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized - Invalid secret key" },
          "409": { "description": "Conflict - Super admin already exists" }
        },
        "summary": "Create the initial super admin (one-time operation)",
        "tags": ["auth"]
      }
    },
    "/auth/register-with-organization": {
      "post": {
        "operationId": "AuthController_registerWithOrganization",
        "parameters": [
          {
            "name": "invitationToken",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "201": { "description": "User registered successfully" },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized - Invalid invitation" }
        },
        "summary": "Register a new user with organization",
        "tags": ["auth"]
      }
    },
...
  },
  "info": {
    "title": "Multi-Sector Management API",
    "description": "API for managing organizations across multiple sectors",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "admin@example.com" },
          "password": { "type": "string", "example": "password123" }
        },
        "required": ["email", "password"]
      },
      "UserDepartmentDto": {
        "type": "object",
        "properties": {
          "departmentId": { "type": "string", "description": "Department ID" },
          "position": {
            "type": "string",
            "description": "User position in department"
          },
          "isHead": {
            "type": "boolean",
            "description": "Whether user is head of department"
          }
        },
        "required": ["departmentId"]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "description": "User email" },
          "password": { "type": "string", "description": "User password" },
          "firstName": { "type": "string", "description": "User first name" },
          "lastName": { "type": "string", "description": "User last name" },
          "isActive": {
            "type": "boolean",
            "description": "User active status",
            "default": true
          },
          "organizationId": {
            "type": "string",
            "description": "Organization ID"
          },
          "roles": {
            "description": "Role IDs to assign",
            "type": "array",
            "items": { "type": "string" }
          },
          "departments": {
            "description": "Departments to assign",
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserDepartmentDto" }
          },
          "sectorProfile": {
            "type": "object",
            "description": "Sector-specific user profile data"
          },
          "address": { "type": "object", "description": "User address" }
        },
        "required": ["email", "password", "firstName", "lastName"]
      },
...
    }
  }
}